+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(bs4Dash)
+
+test_check("bs4Dash")